Add client tools subpackage.
