+++ src/cli.ts
@@ -8,7 +8,8 @@ import { createEditorEventSource, registerEditorListeners } from './chat/editor'
 import { handler } from './chat/handler'
 import { loadHistory } from './chat/history'
 import { ContextStateManager, createContextState } from './context/state'
-import { createClient, registerTools } from './mcp/client/client'
+import { createClient } from './mcp/client/client'
+import { registerTools } from './mcp/client/tools/tools'
 import { Provider } from './providers/provider'
 import { createProvider, modelNames } from './providers/providers'
 import { createInterruptHandler, InterruptHandlerOptions } from './util/interrupts/interrupts'
