+++ src/mcp/server/tools/tools.ts
@@ -1,3 +1,4 @@
+import { ExecutionContext } from './context'
 import { longTask } from './editor/long-task'
 import { editorNotice } from './editor/notice'
 import { Tool } from './tool'
@@ -11,4 +12,20 @@ export function findTool(name: string): Tool {
     }
 
     return tool
+}
+
+export function executeTool(context: ExecutionContext, name: string, args: any): any {
+    try {
+        return findTool(name).execute(context, args)
+    } catch (error: any) {
+        return {
+            content: [
+                {
+                    type: 'text',
+                    text: JSON.stringify({ error: error instanceof Error ? error.message : String(error) }),
+                },
+            ],
+            isError: true,
+        }
+    }
 }
